requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=22.0.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.7.0

# ML & NLP (optional — install for NER support)
# transformers>=4.47.0
# torch>=2.5.0
# spacy>=3.8.0

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
asyncpg>=0.30.0
# alembic>=1.14.0        # for migrations

# Document Parsing
pypdf2>=3.0.1
python-docx>=1.1.2
chardet>=5.2.0

# Reporting
reportlab>=4.2.0
jinja2>=3.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=44.0.0

# Alerting
httpx>=0.28.0
aiosmtplib>=3.0.0

# Streaming (optional)
# aiokafka>=0.12.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0

# Utilities
pyyaml>=6.0.2
python-dotenv>=1.0.1
aiofiles>=24.1.0
